requirements.txt
streamlit
faiss-cpu
google-generativeai
langgraph
langchain-core
llama-index-core
llama-index-embeddings-huggingface
llama-index-vector-stores-faiss
sentence-transformers
